Update on region coordinates
I made some updates regarding region coordinates. Changes were made for generation and for generating the sectioned domain for finding contacts.

During generation, I found that the overlapping was not done properly for the case when a layer was considered. Also, a casting was missing which seems to not have affected the execution of the code. So I corrected how the overlapping flags are updated.

During generation of the sectioned domain for finding contacts. calculations were correct but different from the ones used during generation. The difference in calculating the number of subregions resulting in having more subregions. So I modified this to match the way it is calculated during generation.

Also, I refined the discretization of the GNP to find the subregions it belongs to. In this way it matches the size of the discretization done during generation.