Ignoring small interpenetrations
I found that the penetration depth (PD) is sometimes too small, in the order of 1 nm or less. I doubt this will affect much the Abaqus or resistor network simulations. So I defined a variable for a PD cutoff. Below this PD cutoff interpenetrations are ignored. For now, this cutoff is hardcoded and set to 1 nm. However, the code can easily be modified to include the PD cutoff as an input parameter.