Not all GNP boundary points may be ignored
I made a change in which if a GNP point was at a boundary were voltages were not applied, then this point should be ignored. So, I added a funtion to the Hoshen_Kopelman class to remove such points at boundaries that were not relevant to the problem.

However, I found that there could be a junction at a boundary point (more lilkely after reconstructing a GNP). If one of the junction points belonged to a non-relevant boundary, then the point was deleted but the reference to the point was still there in the junction object. This caused an error.

Thus, I added flags to GNP boundary points. These flags are true if the points is used in a junction and false otherwise. Using these falgs I can avoid deleting points at non-relevant boundaries but that take part of a junction.

Also, I added some comments for debugging.