Corrected application of Abaqus displacements for mixed fillers
I noticed that the application of displacements from Abaqus was not working properly in the case of mixed fillers. The way the if-statements were implemented (as an if-else), displacements were only applied to CNTs. So I split the if-statement into two independent statements and now the displacements may be applied to both CNTs and GNPs when mixed fillers are present.

I also added some lines to measure the time to apply displacements to CNTs and the time for GNPs. This is only printed to file in the case of mixed fillers. It may be useful to know the time it takes to apply displacments to each type of filler.