Ignoring junctons with invalid nodes
I found that some CNT points are removed and I'm not sure where. This results in some junctions haven a CNT point that is not associated with any node. A workaround this issue is to first check if the points have a node associated and if so then the junction is added. Other wise no junction is added. This is the same I did for GNP-GNP junctions and now I implement it for CNT-CNT and mixed junctions.

Also added some comments for debugging and increased the iterations for the CG.