New application to read displacements from files
Added a new application, where the network is read from files and the displacements from Aabqus are read from binary files instead of the odb file. This application is used after using the Extractor program so that the displacements at each frame are available in binary files.

I also modified the file with the main to added the new class and the Input_Reader class to read the number of frames from the input file. The make file was also updated to compile the new class. The input file was modified to add the new application and indicate where the number of frames from Abaqus is to be indicated.