Completing simplex of size 3
When adding displacements to GNP vertices, the four vertices of their faces might not remain in a plane. This results in having simplices of size three when finding junction points between GNPS, which usually does not happen. So, I need to consider this case when displacements are applied only. Otherwise it is still an error.

So I created a flag to dcide whether or not allow simplices of size 3. I also added a function to complete the simplex so that it has size 4 and the flow of the code continues as usual. This might not be accurate in some cases, but I will leave it as a first approximation.

I also made these changes:

1) Commented a line that sent a message to the output file that I forgot to comment.

2) Added some variables to the initialization function for the input parameters.